add(ContactsIndexPage): add dismissable widget for removing friend
