+++ lib/contacts/presentation/view/contacts_export.dart
@@ -1,4 +1,4 @@
 export 'blocked_page.dart';
-export 'contacts_index_page.dart';
+export 'contacts_index_page/contacts_index_page.dart';
 export 'friend_requests_page.dart';
 export 'outgoing_requests_page.dart';
