+++ lib/contacts/presentation/view/contacts_index_page/controller/contacts_list_event.dart
@@ -1,0 +1,19 @@
+part of 'contacts_list_bloc.dart';
+
+@immutable
+abstract class ContactsListEvent extends Equatable {
+  const ContactsListEvent();
+  @override
+  List<Object?> get props => [];
+}
+
+class GetContactsList extends ContactsListEvent {}
+
+class RemoveContact extends ContactsListEvent {
+  const RemoveContact(this.user);
+
+  final User user;
+
+  @override
+  List<Object> get props => [];
+}
