+++ lib/contacts/presentation/view/contacts_index_page/controller/contacts_list_state.dart
@@ -1,0 +1,23 @@
+// ignore_for_file: must_be_immutable
+
+part of 'contacts_list_bloc.dart';
+
+@immutable
+abstract class ContactsListState extends Equatable {
+  late List<User> contactsList;
+  @override
+  List<Object> get props => contactsList;
+}
+
+class ContactsListInitial extends ContactsListState {}
+
+class ContactsListLoadInProgress extends ContactsListState {}
+
+class ContactsListLoadSuccess extends ContactsListState {
+  ContactsListLoadSuccess(this.newContactsList);
+  final List<User> newContactsList;
+  @override
+  List<User> get contactsList => newContactsList;
+}
+
+class ContactsLoadFailure extends ContactsListState {}
