+++ lib/utils/ui_utils/bottom_navigation_bar.dart
@@ -3,8 +3,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:ui_library/ui_library_export.dart';
 import 'package:uplink/chat/chat_export.dart';
-import 'package:uplink/contacts/presentation/controller/friend_bloc.dart';
-import 'package:uplink/contacts/presentation/view/contacts_index_page.dart';
+import 'package:uplink/contacts/presentation/view/contacts_index_page/contacts_index_page.dart';
+import 'package:uplink/contacts/presentation/view/contacts_index_page/controller/contacts_list_bloc.dart';
 import 'package:uplink/file/presentation/controller/item_list_bloc.dart';
 import 'package:uplink/file/presentation/view/file_index_page.dart';
 import 'package:uplink/profile/presentation/controller/current_user_bloc.dart';
@@ -39,7 +39,7 @@ class _MainBottomNavigationBarState extends State<MainBottomNavigationBar> {
       if (_currentIndex == 1) {
         GetIt.I.get<ItemListBloc>().add(GetItemList());
       } else if (_currentIndex == 2) {
-        GetIt.I.get<FriendBloc>().add(ListFriendsStarted());
+        GetIt.I.get<ContactsListBloc>().add(GetContactsList());
       }
     });
   }
